# [IMP] filter: fix horrible performances with huge data filters