--- a/src/components/filters/filter_menu_value_list/filter_menu_value_list.xml
+++ b/src/components/filters/filter_menu_value_list/filter_menu_value_list.xml
@@ -9,7 +9,7 @@
         class="w-100 o-input my-2"
         t-ref="filterMenuSearchBar"
         type="text"
-        t-model="state.textFilter"
+        t-on-input="updateSearch"
         placeholder="Search..."
         t-on-keydown="onKeyDown"
       />
@@ -22,7 +22,7 @@
       t-ref="filterValueList"
       t-on-click="this.clearScrolledToValue"
       t-on-scroll="this.clearScrolledToValue">
-      <t t-foreach="displayedValues" t-as="value" t-key="value.string">
+      <t t-foreach="state.displayedValues" t-as="value" t-key="value.string">
         <FilterMenuValueItem
           onClick="() => this.checkValue(value)"
           onMouseMove="() => this.onMouseMove(value)"
@@ -33,7 +33,13 @@
         />
       </t>
       <div
-        t-if="displayedValues.length === 0"
+        t-if="state.hasMoreValues"
+        class="o-filter-load-more o-button-link d-flex justify-content-center py-1"
+        t-on-click="this.loadMoreValues">
+        Load more...
+      </div>
+      <div
+        t-if="state.displayedValues.length === 0"
         class="o-filter-menu-no-values d-flex align-items-center justify-content-center w-100 h-100 ">
         No results
       </div>