--- a/src/plugins/ui_stateful/filter_evaluation.ts
+++ b/src/plugins/ui_stateful/filter_evaluation.ts
@@ -181,9 +181,10 @@ export class FilterEvaluationPlugin extends UIPlugin {
       if (filterValue.filterType === "values") {
         const filteredValues = filterValue.hiddenValues?.map(toLowerCase);
         if (!filteredValues) continue;
+        const filteredValuesSet = new Set(filteredValues);
         for (let row = filteredZone.top; row <= filteredZone.bottom; row++) {
           const value = this.getCellValueAsString(sheetId, filter.col, row);
-          if (filteredValues.includes(value)) {
+          if (filteredValuesSet.has(value)) {
             hiddenRows.add(row);
           }
         }