… tests/table/__snapshots__/filter_menu_component.test.ts.snap