@robodoo 2025-05-05
This PR targets the un-managed branch odoo/o-spreadsheet:master-data-filter-rule-adrm, it needs to be retargeted before it can be merged.