@robodoo 2025-05-12
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/o-spreadsheet/pull/6286.png)](https://mergebot.odoo.com/odoo/o-spreadsheet/pull/6286)